+++ Changelog.md
@@ -12,6 +12,7 @@ OpenCore Changelog
 - Added OEM preservation for certain Apple SMBIOS tables
 - Fixed switching to graphics mode when entering OpenCanopy
 - Fixed installing Apple FB Info protocol when no GOP exists
+- Fixed abort timeout sound in OpenCanopy on key press
 
 #### v0.6.6
 - Added keyboard and pointer entry scroll support in OpenCanopy